+++ StudentAndCourses/Classes/StudentRepository.cs
@@ -7,6 +7,8 @@ using System.Diagnostics.Eventing.Reader;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Windows.Forms;
+using System.Windows;
 using Dapper;
 
 namespace StudentAndCourses.Classes
@@ -14,16 +16,10 @@ namespace StudentAndCourses.Classes
     internal class StudentRepository
     {
         private readonly IDbConnection dbConnection;
-        private SqlConnection connection;
 
         public StudentRepository(string con2)
         {
             dbConnection = new SqlConnection(con2);
-        }
-
-        public StudentRepository(SqlConnection connection)
-        {
-            this.connection = connection;
         }
 
         public List<Student> SelectedStudents()
@@ -32,8 +28,20 @@ namespace StudentAndCourses.Classes
                 return students;
         }
 
+        public bool IsStudentExist(string name, string surname)
+        {
+            string sqlQuery = "SELECT COUNT(*) FROM Students WHERE SName = @Name AND SSurname = @Surname";
+            int count = dbConnection.QuerySingle<int>(sqlQuery, new { Name = name, Surname = surname });
+            return count > 0;
+        }
+
         public void AddStudent(string name, string surname, int age)
         {
+            if (IsStudentExist(name, surname))
+            {
+                System.Windows.MessageBox.Show("Студент із таким ім'ям та прізвищем вже існує", "Увага", MessageBoxButton.OK, (MessageBoxImage)MessageBoxIcon.Warning);
+            }
+
             dbConnection.Execute("INSERT INTO Students (SName, SSurname, SAge) VALUES (@Name, @Surname, @Age)",
                                   new { Name = name, Surname = surname, Age = age });
         }
@@ -43,12 +51,27 @@ namespace StudentAndCourses.Classes
             string sql = "UPDATE Students SET SName = @Name, SSurname = @Surname, SAge = @Age WHERE id_students = @id_students";
 
             dbConnection.Execute(sql, new { Name = name, Surname = surname, Age = age, id_students = id });
+            System.Windows.MessageBox.Show("Дані успішно оновлені", "Успішно", MessageBoxButton.OK, (MessageBoxImage)MessageBoxIcon.Information);
         }
 
         public void DeleteStudent(int id)
         {
-            string sqlQuery = "DELETE FROM Students WHERE id_students = @id_students";
-            dbConnection.Execute(sqlQuery, new { id_students = id });
+            try
+            {
+                if (System.Windows.MessageBox.Show("Ви точно хочете видалити студента?", "Попередження", MessageBoxButton.YesNo, (MessageBoxImage)MessageBoxIcon.Question) == MessageBoxResult.Yes)
+                {
+                    string sqlQuery = "DELETE FROM Students WHERE id_students = @id_students";
+                    dbConnection.Execute(sqlQuery, new { id_students = id });
+                    System.Windows.MessageBox.Show("Успішно видалено", "Успішно", MessageBoxButton.OK, (MessageBoxImage)MessageBoxIcon.Information);
+                }
+            }
+            catch (SqlException ex)
+            {
+                if (ex.Number == 547)
+                {
+                    System.Windows.MessageBox.Show("Студент не може бути видалений, тому що є пов'язані записи у списку призначених курсів", "Попередження", MessageBoxButton.OK, (MessageBoxImage)MessageBoxIcon.Warning);
+                }
+            }
         }
     }
 }
